force_calibration: split calibration into separate files
